test: Fix tests, add new functions
